requirements.txt
fastapi[all]
langchain
langchain-community
langchain-core
langchain-openai 
line-bot-sdk
pinecone-client==2.2.1
tiktoken
uvicorn